requirements.dev.txt
Jinja2==2.6
Pygments==1.5
Sphinx==1.1.3
Werkzeug==0.8.3
django-debug-toolbar==0.9.4
django-devserver==0.4.0
django-extensions==1.0.1
